Split the media-session browser tier by file
node:test applies --test-timeout to the file-level test as well as each
subtest, so adding a case that plays real media pushed the file past 30s
even though every subtest passed. Multi-frame coverage moves to its own
file over a shared page harness, and the tier's budget goes to 90s: for a
tier driving real playback the timeout is a hang guard, not a target.
